requirements.txt
GDAL==3.11.3
pillow==11.3.0
numpy==2.3.2
